route guards/ started implementing edit

+++ client/src/components/catalog/PCConfigurationInfo.jsx
@@ -1,4 +1,4 @@
-import { Link, useNavigate, useParams } from "react-router";
+import { useNavigate, useParams } from "react-router";
 import useUserAuth from "../../hooks/useUserAuth";
 import { useDeleteConfiguration, useOneConfiguration } from "../../api/PCConfigurationsApi";
 
@@ -10,21 +10,25 @@ export default function PCConfigurationInfo() {
     const { deleteConfiguration } = useDeleteConfiguration();
 
     const configurationDeleteHandler = async () => {
-        // const hasConfirm = confirm(`Are you sure you want to delete ${configuration.name}?`);
+        const hasConfirm = confirm(`Are you sure you want to delete ${configuration.name}?`);
 
-        // if (!hasConfirm) {
-        //     return;
-        // }
+        if (!hasConfirm) {
+            return;
+        }
 
         await deleteConfiguration(configurationId);
 
         navigate('/configurations');
     }
 
+    const configurationEditHandler = () => {
+        navigate(`/configurations/${configurationId}/edit`);
+    }
+
     const isOwner = userId === configuration._ownerId;
 
     return (
-        <div class="configuration-details">
+        <div className="configuration-details">
             <h2>Configuration Info</h2>
             <div className="image-container">
                 <img src={configuration.image} alt="Configuration Image" />
@@ -42,12 +46,9 @@ export default function PCConfigurationInfo() {
 
             {isOwner && (
                 <>
-                    <Link>
-                        <button className="details-btn">Edit</button>
-                    </Link>
-                    <Link>
-                        <button onClick={configurationDeleteHandler} className="details-btn">Delete</button>
-                    </Link>
+                    <button onClick={configurationEditHandler} className="details-btn">Edit</button>
+
+                    <button onClick={configurationDeleteHandler} className="details-btn">Delete</button>
                 </>
             )}
         </div >
